+++ src/pages/login/index.tsx
@@ -1,122 +1,124 @@
+import type { NextPage } from 'next';
 import { useRouter } from 'next/router';
-import { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 
+import AppleIcon from '@assets/apple.svg';
+import KakaoIcon from '@assets/kakao.svg';
 import Mascot from '@assets/mascot.svg';
-import LoginLayout from '@components/Layout/LoginLayout';
+import NaverIcon from '@assets/naver.svg';
+import PlainOMOLetter from '@assets/omo-letter.svg';
 
-type ErrorType = 'default' | 'duplicate' | 'usable';
-
-const errorStatusMsg = {
-  default: '2자 이상 입력해주세요',
-  duplicate: '중복된 닉네임입니다',
-  usable: '사용가능한 닉네임입니다',
-};
-
-const errorStatusColor = {
-  default: '#000',
-  duplicate: '#FF473B',
-  usable: '#2334CF',
-};
-
-const Index = () => {
+const Login: NextPage = () => {
   const router = useRouter();
-  const [errorStatus] = useState<ErrorType>('usable');
 
-  const onClickNextButton = () => {
-    if (errorStatus !== 'usable') return;
-    router.push('/login/profile');
+  const onClickOAuthButton = (e: React.MouseEvent<HTMLDivElement>) => {
+    if (!(e.target instanceof Element)) return;
+
+    if (e.target.closest('.oauth-button')) {
+      // TODO: 추후에 플랫폼에 맞게 처리
+      router.push('/login/nickname');
+    }
   };
 
   return (
-    <LoginLayout>
+    <HomePage>
       <Content>
-        <Mascot />
-        <div className="welcome-letter">어서오세요!</div>
-        <div className="welcome-sub-letter">닉네임을 정해주세요 (최대 10자)</div>
-        <div className="nickname-input">
-          <input type="text" placeholder="닉네임 입력" maxLength={10} />
-          <ErrorNotifySpan errorStatus={errorStatus}>{errorStatusMsg[errorStatus]}</ErrorNotifySpan>
-        </div>
+        <LogoSection>
+          <Mascot />
+          <LogoLetter />
+          <SubLetter>오마카세의</SubLetter>
+          <SubLetter>모든것</SubLetter>
+        </LogoSection>
+        <SNSSection>
+          <span>SNS 간편로그인</span>
+          <div className="oauth-button-wrapper" onClick={onClickOAuthButton}>
+            <div className="oauth-button naver">
+              <NaverIcon />
+              <div>Naver</div>
+            </div>
+            <div className="oauth-button kakao">
+              <KakaoIcon />
+              <div>Kakao</div>
+            </div>
+            <div className="oauth-button apple">
+              <AppleIcon />
+              <div>Apple</div>
+            </div>
+          </div>
+        </SNSSection>
       </Content>
-      <NextPageButton errorStatus={errorStatus} onClick={onClickNextButton}>
-        다음
-      </NextPageButton>
-    </LoginLayout>
+    </HomePage>
   );
 };
 
-export default Index;
+export default Login;
 
-const Content = styled.div`
-  box-sizing: border-box;
+const HomePage = styled.main`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: #2334cf;
+`;
+
+const LogoSection = styled.section`
+  height: fit-content;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  svg:first-child {
+    transform: scale(0.8);
+  }
+`;
+
+const SNSSection = styled.section`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+
   width: 100%;
-  height: 310px;
-  margin-top: 6rem;
-  padding: 0 20px;
+  height: 123px;
+  font-size: 12px;
+  color: #fff;
 
-  svg {
-    display: block;
-    margin: 0 auto;
-  }
-
-  .welcome-letter {
-    font-weight: bold;
-    font-size: 30px;
-    line-height: 36px;
-    margin: 29px 0 6px 0;
-  }
-
-  .welcome-sub-letter {
-    font-weight: 600;
-    font-size: 14px;
-    margin-bottom: 40px;
-  }
-
-  .nickname-input {
-    position: relative;
+  .oauth-button-wrapper {
     width: 100%;
+    display: flex;
+    justify-content: space-between;
   }
 
-  input {
-    all: unset;
-    width: 100%;
-    height: 29px;
-    border-bottom: 1px solid #000;
+  .oauth-button {
+    width: 60px;
+    display: inline;
+    cursor: pointer;
 
-    color: #000;
-
-    ::placeholder {
-      color: #b8b8b8;
+    div {
+      width: fit-content;
+      margin: 7px auto 0;
     }
   }
 `;
 
-const ErrorNotifySpan = styled.span<{ errorStatus: ErrorType }>`
-  position: absolute;
-  bottom: 4px;
-  right: 0;
-  line-height: 18px;
-  font-size: 12px;
-
-  color: ${({ errorStatus }) => errorStatusColor[errorStatus]};
+const Content = styled.div`
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  width: 225px;
+  height: 500px;
 `;
 
-const NextPageButton = styled.button<{ errorStatus: ErrorType }>`
-  border: none;
-  position: absolute;
-  left: 20px;
-  bottom: 3rem;
-  width: calc(100% - 40px);
-  height: 48px;
+const LogoLetter = styled(PlainOMOLetter)`
+  margin: 5px 0 15px;
+`;
 
-  margin: 0 auto;
-  border-radius: 8px;
-
-  background-color: ${({ errorStatus }) => (errorStatus === 'usable' ? '#2334CF' : '#d7d8dd')};
+const SubLetter = styled.div`
+  ${({ theme }) => theme.fonts.bold};
+  letter-spacing: -0.02em;
+  line-height: 18px;
   color: #fff;
-
-  font-size: 18px;
-  font-weight: 700;
-  cursor: ${({ errorStatus }) => (errorStatus === 'usable' ? 'pointer' : 'default')};
 `;
